test: Refactor test codes
